# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  response.PaginatedData:
    properties:
      items: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  service.ComboPurchaseResponse:
    properties:
      activated:
        type: boolean
      activated_at:
        type: string
      combo_id:
        type: string
      created_at:
        type: string
      current_session_id:
        type: string
      expires_at:
        type: string
      id:
        type: string
      member_id:
        type: string
      payment_method:
        type: string
      price:
        type: integer
      remaining_minutes:
        type: integer
    type: object
...
  service.CreateMemberRequest:
    properties:
      date_of_birth:
        type: string
      email:
        type: string
      full_name:
        type: string
      id_card_number:
        type: string
      notes:
        type: string
      password:
        type: string
      phone:
        type: string
      store_id:
        type: string
      tier_id:
        type: string
    type: object
...
  service.MemberResponse:
    properties:
      avatar_url:
        type: string
      balance:
        type: integer
      bonus_balance:
        type: integer
      code:
        type: string
      created_at:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      id_card_image_url:
        type: string
      id_card_number:
        type: string
      is_active:
        type: boolean
      last_visit_at:
        type: string
      notes:
        type: string
      parent_consent_file_url:
        type: string
      phone:
        type: string
      store_id:
        type: string
      tier:
        $ref: '#/definitions/service.TierResponse'
      tier_id:
        type: string
      total_played_hours:
        type: integer
      total_spent:
        type: integer
      updated_at:
        type: string
    type: object
  service.MemberTransactionResponse:
    properties:
      amount:
        type: integer
      balance_after:
        type: integer
      balance_before:
        type: integer
      bonus_after:
        type: integer
      bonus_before:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      member_id:
        type: string
      payment_method:
        type: string
      reference_id:
        type: string
      store_id:
        type: string
      transaction_type:
        type: string
    type: object
...
  service.RefundRequest:
    properties:
      amount:
        type: integer
      description:
        type: string
      is_bonus:
        type: boolean
    type: object
...
  service.SessionResponse:
    properties:
      combo_id:
        type: string
      combo_type:
        type: string
      created_at:
        type: string
      duration_minutes:
        type: integer
      ended_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_overnight:
        type: boolean
      machine_id:
        type: string
      member_id:
        type: string
      remaining_minutes:
        type: integer
      slot_end:
        type: string
      started_at:
        type: string
      store_id:
        type: string
      total_cost:
        type: integer
    type: object
...
  service.TierResponse:
    properties:
      created_at:
        type: string
      discount_percent:
        type: number
      id:
        type: string
      min_spent:
        type: integer
      name:
        type: string
    type: object
...
  service.TopupRequest:
    properties:
      amount:
        type: integer
      description:
        type: string
      payment_method:
        type: string
    type: object
...
  service.UpdateMemberRequest:
    properties:
      avatar_url:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      full_name:
        type: string
      id_card_number:
        type: string
      is_active:
        type: boolean
      notes:
        type: string
      password:
        type: string
      phone:
        type: string
      tier_id:
        type: string
    type: object
...
host: localhost:20800
info:
  contact:
    email: support@vnet.net
    name: VNET Support
  description: Hệ thống Quản lý Phòng Game Toàn Diện
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://vnet.net/terms
  title: VNET Core API
  version: "1.0"
paths:
...
  /members:
    get:
      consumes:
      - application/json
      description: Get a paginated list of members
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      - description: Search keyword
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PaginatedData'
                  - properties:
                      data:
                        items:
                          $ref: '#/definitions/service.MemberResponse'
                        type: array
                    type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List members
      tags:
      - Members
    post:
      consumes:
      - application/json
      description: Create a new member record
      parameters:
      - description: Member details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.CreateMemberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.MemberResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a new member
      tags:
      - Members
  /members/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a member by ID
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete a member
      tags:
      - Members
    get:
      consumes:
      - application/json
      description: Get a single member by their ID
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.MemberResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get member by ID
      tags:
      - Members
    put:
      consumes:
      - application/json
      description: Update an existing member's details
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated member details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.UpdateMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.MemberResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update a member
      tags:
      - Members
  /members/{id}/combos:
    get:
      consumes:
      - application/json
      description: Get a paginated list of combo purchases for a member
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PaginatedData'
                  - properties:
                      data:
                        items:
                          $ref: '#/definitions/service.ComboPurchaseResponse'
                        type: array
                    type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get member combo purchases
      tags:
      - Members
  /members/{id}/refund:
    post:
      consumes:
      - application/json
      description: Refund funds from a member's account
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: string
      - description: Refund details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.RefundRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.MemberResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Refund member balance
      tags:
      - Members
  /members/{id}/sessions:
    get:
      consumes:
      - application/json
      description: Get a paginated list of play sessions for a member
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PaginatedData'
                  - properties:
                      data:
                        items:
                          $ref: '#/definitions/service.SessionResponse'
                        type: array
                    type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get member sessions
      tags:
      - Members
  /members/{id}/topup:
    post:
      consumes:
      - application/json
      description: Add funds to a member's account
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: string
      - description: Topup details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.TopupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.MemberResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Topup member balance
      tags:
      - Members
  /members/{id}/transactions:
    get:
      consumes:
      - application/json
      description: Get a paginated list of transactions for a member
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PaginatedData'
                  - properties:
                      data:
                        items:
                          $ref: '#/definitions/service.MemberTransactionResponse'
                        type: array
                    type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get member transactions
      tags:
      - Members
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"